Split Helpers/EventMethods.cs into Interop partials and extract LoadLibraryFlags
