+++ src/EventLogExpert.Eventing/Helpers/PathType.cs
@@ -1,0 +1,10 @@
+// // Copyright (c) Microsoft Corporation.
+// // Licensed under the MIT License.
+
+namespace EventLogExpert.Eventing.Helpers;
+
+public enum PathType
+{
+    LogName = 1,
+    FilePath
+}
